+++ restiny/data/sql/6_add_editor_indent_to_settings.sql
@@ -1,0 +1,2 @@
+ALTER TABLE settings
+  ADD editor_indent INTEGER NOT NULL DEFAULT 2