+++ restiny/ui/app.py
@@ -1,5 +1,6 @@
 import asyncio
 import json
+import mimetypes
 from collections.abc import Iterable
 from http import HTTPStatus
 
@@ -16,6 +17,7 @@ from textual.widgets import Footer, Header
 
 from restiny.__about__ import __version__
 from restiny.assets import STYLE_TCSS
+from restiny.consts import DOWNLOADS_DIR
 from restiny.data.db import DBManager
 from restiny.data.repos import (
     EnvironmentsSQLRepo,
@@ -37,17 +39,14 @@ from restiny.ui import (
     TopBarArea,
     URLArea,
 )
-from restiny.ui.screens.environments_screen import EnvironmentsScreen
-from restiny.ui.screens.openapi_spec_import_screen import (
+from restiny.ui.screens import (
+    EnvironmentsScreen,
     OpenapiSpecImportScreen,
-)
-from restiny.ui.screens.postman_collection_import_screen import (
     PostmanCollectionImportScreen,
-)
-from restiny.ui.screens.postman_environment_import_screen import (
     PostmanEnvironmentImportScreen,
+    SettingsScreen,
 )
-from restiny.ui.screens.settings_screen import SettingsScreen
+from restiny.utils import is_textual_mimetype
 from restiny.widgets.custom_text_area import CustomTextArea
 
 
@@ -119,8 +118,6 @@ class RESTinyApp(App, inherit_bindings=False):
         self._last_focused_maximizable_area: Widget | None = None
         self._selected_request: Request | None = None
         self._request_id_to_response: dict[int, httpx.Response] = {}
-        self._http_client: httpx.AsyncClient | None = None
-        self._http_client_config: tuple | None = None
         self._cookies: httpx.Cookies = httpx.Cookies()
 
     def compose(self) -> ComposeResult:
@@ -321,6 +318,12 @@ class RESTinyApp(App, inherit_bindings=False):
     def _on_send_request(self, message: URLArea.SendRequest) -> None:
         self._active_request_task = asyncio.create_task(self._send_request())
 
+    @on(URLArea.DownloadResponse)
+    def _on_download_response(self, message: URLArea.DownloadResponse) -> None:
+        self._active_request_task = asyncio.create_task(
+            self._send_request(download=True)
+        )
+
     @on(URLArea.CancelRequest)
     def _on_cancel_request(self, message: URLArea.CancelRequest) -> None:
         if self._active_request_task and not self._active_request_task.done():
@@ -356,11 +359,18 @@ class RESTinyApp(App, inherit_bindings=False):
 
     def _apply_settings(self) -> None:
         settings = self.settings_repo.get().data
+
         self.call_later(lambda: setattr(self, 'theme', settings.theme))
+
         for text_area in self.query(CustomTextArea):
             self.call_later(
                 lambda ta=text_area: setattr(
                     ta, 'theme', settings.editor_theme
+                )
+            )
+            self.call_later(
+                lambda ta=text_area: setattr(
+                    ta, 'indent_width', settings.editor_indent
                 )
             )
 
@@ -591,7 +601,7 @@ class RESTinyApp(App, inherit_bindings=False):
             request.options.attach_cookies
         )
 
-    async def _send_request(self) -> None:
+    async def _send_request(self, download: bool = False) -> None:
         self.response_area.clear()
         self.response_area.loading = True
         self.url_area.request_pending = True
@@ -615,9 +625,43 @@ class RESTinyApp(App, inherit_bindings=False):
 
             if request.options.attach_cookies:
                 self._cookies.extract_cookies(response)
+
+            if download:
+                content_disposition = response.headers.get(
+                    'content-disposition'
+                )
+                if content_disposition and 'filename=' in content_disposition:
+                    filename = content_disposition.split('filename=')[
+                        -1
+                    ].strip('"')
+                else:
+                    filename = (
+                        response.url.path.removeprefix('/').removesuffix('/')
+                        or 'response'
+                    )
+                filename = filename.rsplit('.', 1)[0]
+
+                content_type = response.headers.get('content-type')
+                if content_type:
+                    filesuffix = (
+                        mimetypes.guess_extension(content_type.split(';')[0])
+                        or '.bin'
+                    )
+                else:
+                    filesuffix = '.bin'
+
+                download_file = DOWNLOADS_DIR / f'{filename}{filesuffix}'
+                counter = 1
+                while download_file.exists():
+                    download_file = (
+                        DOWNLOADS_DIR / f'{filename}({counter}){filesuffix}'
+                    )
+                    counter += 1
+                download_file.write_bytes(response.content)
+                self.notify(f'Download file {download_file}')
+
             self._display_response(response=response)
             self.response_area.is_showing_response = True
-
             self._request_id_to_response[request.id] = response
 
         except httpx.RequestError as error:
@@ -656,13 +700,29 @@ class RESTinyApp(App, inherit_bindings=False):
             header_key: header_value
             for header_key, header_value in response.headers.multi_items()
         }
+
+        content_type = response.headers.get('Content-Type', '')
+        mimetype = content_type.split(';', 1)[0].strip().lower()
+
         self.response_area.body_raw_language = (
-            content_type_to_body_language.get(
-                response.headers.get('Content-Type'), BodyRawLanguage.PLAIN
-            )
+            content_type_to_body_language.get(mimetype, BodyRawLanguage.PLAIN)
         )
         if self.response_area.body_raw_language == BodyRawLanguage.JSON:
-            indented_body_raw = json.dumps(json.loads(response.text), indent=4)
-            self.response_area.body_raw = indented_body_raw
-        else:
+            try:
+                self.response_area.body_raw = json.dumps(
+                    json.loads(response.text),
+                    indent=self.settings_repo.get().data.editor_indent,
+                )
+                return
+            except json.JSONDecodeError:
+                self.notify(
+                    message='Content-Type is JSON, but the response body is not valid JSON',
+                    severity='warning',
+                )
+
+        if is_textual_mimetype(mimetype=mimetype):
             self.response_area.body_raw = response.text
+        else:
+            self.response_area.body_raw = (
+                '[BINARY CONTENT]\nPress "Download" to save'
+            )